requirements.txt
# Core async and networking
websockets>=12.0
httpx>=0.27.0
aiohttp>=3.9.0

# API clients
mistralai>=0.1.0

# Local STT
faster-whisper>=1.0.0

# Local TTS
piper-tts>=1.2.0

# Audio processing
pyaudio>=0.2.14

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
numpy
